Revert "[SPARK-36894][SPARK-37077][PYTHON] Synchronize RDD.toDF annotations with SparkSession and SQLContext .createDataFrame variants."
This reverts commit 855da09f02f3007a2c36e7a738d4dc81fd95569a.
See [this comment](https://github.com/apache/spark/pull/34146#issuecomment-959136935).

Closes #34477 from sarutak/revert-SPARK-37077.

Authored-by: Kousuke Saruta <sarutak@oss.nttdata.com>
